sp_service/requirements.txt
amqp==5.1.1
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
Babel==2.10.3
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
black==22.8.0
bleach==5.0.1
blessed==1.19.1
botocore==1.23.54
cached-property==1.5.2
celery==5.2.7
cement==2.8.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
cred==0.2.3
cryptography==38.0.3
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-extensions==3.2.0
django-rest-framework==0.1.0
django-timezone-field==5.0
djangorestframework==3.13.1
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.4
executing==0.9.1
fastjsonschema==2.16.1
flower==1.1.0
future==0.16.0
gunicorn==20.1.0
holidaycal==0.0.1
humanize==4.2.3
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
jmespath==0.10.0
json5==0.9.8
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyterlab==3.4.4
jupyterlab-pygments==0.2.2
jupyterlab-server==2.15.0
jupyterlab-widgets==1.1.1
kombu==5.2.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclassic==0.4.3
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
notebook-shim==0.1.0
numpy==1.23.1
oauthlib==3.2.1
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
paramiko==2.12.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
psycopg2==2.9.3
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.1.0
redis==4.3.4
requests==2.26.0
requests-oauthlib==1.3.1
semantic-version==2.8.5
Send2Trash==1.8.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
spotipy==2.20.0
sqlparse==0.4.2
stack-data==0.3.0
termcolor==1.1.0
terminado==0.15.0
texttable==1.6.4
tinycss2==1.1.1
tomli==2.0.1
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.10
uWSGI==2.0.21
vine==5.0.0
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.59.0
Werkzeug==2.2.1
widgetsnbextension==3.6.1
wrapt==1.14.1
zipp==3.8.1
